% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergas_spat.R
\name{ergas_spat}
\alias{ergas_spat}
\title{ERGAS Spatial Quality index}
\usage{
ergas_spat(
  original,
  modified,
  pan,
  method,
  rp = mean(res(pan)),
  rms = mean(res(original))
)
}
\arguments{
\item{original}{Raster brick object with the original multispectral bands}

\item{modified}{Raster brick object with the bands obtained by the fusion process}

\item{pan}{Raster layer object with the panchromatic band}

\item{method}{Resampling method, should be ‘"bilinear"’ for bilinear interpolation, or ‘"ngb"’}

\item{rp}{Resolution of the original panchromatic band}

\item{rms}{Resolution of the original multispectral bands}
}
\value{
The value of the quality index
}
\description{
ERGAS Spatial Quality index
}
\examples{
pca <- pca_fusion(mis=mis, pan=pan, method="bilinear", bits=16)
ergas_spat(original=mis, modified=pca, pan=pan, method="bilinear")

}
