% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergas_spec.R
\name{ergas_spec}
\alias{ergas_spec}
\title{ERGAS Spectral Quality index}
\usage{
ergas_spec(
  original,
  modified,
  pan,
  method,
  rp = mean(res(pan)),
  rms = mean(res(original))
)
}
\arguments{
\item{original}{Raster brick object with the original multispectral bands}

\item{modified}{Raster brick object with the bands obtained by the fusion process}

\item{pan}{Raster object with the panchromatic band.}

\item{method}{Resampling method, should be ‘"bilinear"’ for bilinear interpolation, or ‘"ngb"’}

\item{rp}{Resolution of the original panchromatic band}

\item{rms}{Resolution of the original multispectral bands}
}
\value{
Quality index
}
\description{
ERGAS Spectral Quality index
}
\examples{
pca <- pca_fusion(mis=mis, pan=pan, method="bilinear", bits=16)
ergas_spec(original=mis, modified=pca, pan=pan, method="bilinear")

}
