% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix.R
\name{get_matrix}
\alias{get_matrix}
\title{Calculate filtering matrix for High Pass Filter. Internally used by HPF_fusion function}
\usage{
get_matrix(r)
}
\arguments{
\item{r}{A number with the size of the matriz}
}
\value{
A matrix object with the filtering matrix whose size depends on r
}
\description{
Calculate filtering matrix for High Pass Filter. Internally used by HPF_fusion function
}
\examples{
get_matrix(3)

}
