% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf_fusion.R
\name{hpf_fusion}
\alias{hpf_fusion}
\title{High Pass Filter Transformation}
\usage{
hpf_fusion(mis, pan, method = "bilinear", bits)
}
\arguments{
\item{mis}{Raster brick object with the original multispectral bands}

\item{pan}{Raster layer object with the panchromatic band}

\item{method}{Resampling method, should be ‘"bilinear"’ for bilinear interpolation, or ‘"ngb"’}

\item{bits}{Radiometric resolution of the original multispectral bands}
}
\value{
Raster brick object with the bands obtained by the fusion process
}
\description{
High Pass Filter Transformation
}
\examples{
hpf <- hpf_fusion(mis=mis, pan=pan, method="bilinear", bits=16)

}
