% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_fusion.R
\name{pca_fusion}
\alias{pca_fusion}
\title{Principal Components Analysis Transformation}
\usage{
pca_fusion(mis, pan, method, bits, matrix = NULL, mode = 1)
}
\arguments{
\item{mis}{Raster brick object with the original multispectral bands}

\item{pan}{Raster layer object with the panchromatic band}

\item{method}{Resampling method, should be ‘"bilinear"’ for bilinear interpolation, or ‘"ngb"’}

\item{bits}{Radiometric resolution of the original multispectral bands}

\item{matrix}{PCA matrix. It is the pca matrix of variable loadings. The usual working mode is to allow the function calculates it, but it can be introduced by the user.}

\item{mode}{With mode=-1 the first component is multiplied by -1}
}
\value{
Raster brick object with the bands obtained by the fusion process
}
\description{
Principal Components Analysis Transformation
}
\examples{
pca <- pca_fusion(mis=mis, pan=pan, method="bilinear", bits=16)

}
